Implement restart client logic in `AcsAgentClient` class on RPC termination and related unit tests.
To implement this mechanism, we added another thread `restart_client_thread_` running `RestartClient()`. `ReactorReadCallback()` will notify `RestartClient()` when the Read operation of `AcsAgentClientReactor` is not ok, which indicates a stream closure. `RestartClient()` will query the termination status of the RPC, generate a new connection id, stub and reactor, and then register the connection.

I added another mutex, `reactor_mtx_`, in `AcsAgentClient` to protect `reactor_` and all variables related to write. This mutex can keep the client from writing (calling reactor_->AddRequest()) if the registration is neither attempted nor successful.

The minor changes on the AcsAgentClientReactor:
1. `read_callback_` will now take another parameter (bool ok) to indicate if the stream is getting closed.
2. Implement a destructor to ensure that `OnDone()` is called and termination status is consumed before destruction. This is to fix some race condition in testing, when the test destructs the reactor object even before `OnDone()` is called.

Also made some minor changes on the FakeAcsAgentServerReactor and FakeAcsAgentServer to facilitate the restart of the server to close the RPC, to test the restart client behavior.

TODO: the Init(), RegistrationConnection(), and RestartClient() have some overlapping code that can be refactored in a follow up CL to reduce the complexity of review on this CL.
PiperOrigin-RevId: 692216575