Add agent_communication.proto file.
For golang, we directly expose the pb.go file to customer of agent library.

However, for C++ users, the pb.h and pb.cc files compiled by protoc of lower version cannot be compiled by protobuf library of higher version. And the following the convention of exposing public APIs, we decide to expose the raw proto file for C++ users.

PiperOrigin-RevId: 694528039