Copy paste all the proto files that agent_communication.proto depends on.
To compile generated code of proto (i.e., .pb.cc, .pb.h, etc), all the dependent proto files must be compiled as well. And they should be put together in a same directory to avoid the descriptor name mismatch. Ref: https://stackoverflow.com/questions/76490108/importing-a-protobuf-file-from-another-directory-and-building-it-with-cmake

The dependent files, client.proto, field_behavior.proto
, launch_stage.proto , and status.proto are all copied from https://github.com/googleapis/googleapis, with their import directory changed because they are now put in the same directory.

PiperOrigin-RevId: 695393792