Support Regional and staging ACS Endpoint.
Previously, the endpoint is hard coded as "<location>-agentcommunication.googleapis.com:443". For staging ACS server, the endpoint should be "<location>-agentcommunication.sandbox.googleapis.com:443".

Also added a factory method for the ease of user to create client, with just a boolean to indicate whether the instance is regional and the channel id.

NO test is added as the added method is static factory method.

PiperOrigin-RevId: 703186633