Get AgentConnectionId by parsing ACS token.
To construct AgentConnectionId, we need to know the instance's project number, zone, and instance id, all of which can be obtained from parsing the ACS token.

We first try to grab the token from GCE default endpoint. If the token is not in valid format, we fallback to GKE endpoint.

PiperOrigin-RevId: 727074044