Fix deadlock when shutting down client with cancelled stream
The previous logic had a chance of a deadlock. When closing a client, the reactor would still have a thread running that would call ReactorReadCallback() which then would try to relock reactor_mtx_ but it was already locked by the caller.
Fix this by holding the reactor in a temp variable, releasing the mutex and then deleting it.

PiperOrigin-RevId: 769238118